+++ SF3000/SFlib/SF_Access.h
@@ -40,12 +40,12 @@ extern char height_map[256][256];        // Pointer to map [height] data
 extern char poly_map[128][128];          // Pointer to map [polygon] data
 extern char animate[2052];               // Pointer to map [animations] data
 extern char cels4x4[16 * CEL_MAX4];      // Pointer to 4x4 compacted land cels
-extern long graphics_data[19000];        // Pointer to polygon datafiles
+extern int32_t graphics_data[19000];     // Pointer to polygon datafiles
 extern char skyfile[1024];               // Skyfile
 extern mission_data mission;             // Mission data
 extern cel_celdata cel_quad;             // Cel info block
 extern char tangent_table[4100];         // Tangent Data
-extern int32_t cosine_table[2048];          // Cosine table
+extern int32_t cosine_table[2048];       // Cosine table
 extern char animate_poly[1024];          // Texture animations file
 extern game_configuration configuration; // Game configuration
